changed jwtExpirationMs to long

+++ src/main/java/com/tweetapp/security/jwt/JwtUtils.java
@@ -25,7 +25,7 @@ public class JwtUtils {
 	private String jwtSecret;
 
 	@Value("${tweet.app.jwtExpirationMs}")
-	private int jwtExpirationMs;
+	private long jwtExpirationMs;
 
 	public String generateJwtToken(Authentication authentication) {
